+++ .ci/run-repository.sh
@@ -65,7 +65,7 @@ elif [[ "$TASK_TYPE" == "doc" ]]; then
   
   docker cp opensearch-py-ml-doc-runner:/code/opensearch-py-ml/docs/build/ ./docs/
   docker rm opensearch-py-ml-doc-runner
-elif [[ "$TASK_TYPE" == "SentenceTransformerTrace" || "$TASK_TYPE" == "SparseTrace" || "$TASK_TYPE" == "SparseTokenizerTrace" || "$TASK_TYPE" == "SemanticHighlighterTrace" ]]; then
+elif [[ "$TASK_TYPE" == "SentenceTransformerTrace" || "$TASK_TYPE" == "SparseTrace" || "$TASK_TYPE" == "SparseTokenizerTrace" || "$TASK_TYPE" == "SemanticHighlighterTrace" || "$TASK_TYPE" == "MetricsCorrelationTrace"]]; then
   # Set up OpenSearch cluster & Run model autotracing (Invoked by model_uploader.yml workflow)
   echo -e "\033[34;1mINFO:\033[0m MODEL_ID: ${MODEL_ID}\033[0m"
   echo -e "\033[34;1mINFO:\033[0m MODEL_VERSION: ${MODEL_VERSION}\033[0m"
@@ -91,6 +91,9 @@ elif [[ "$TASK_TYPE" == "SentenceTransformerTrace" || "$TASK_TYPE" == "SparseTra
       EXTRA_ARGS=( -t )
   elif [[ "$TASK_TYPE" == "SemanticHighlighterTrace" ]]; then
       NOX_TRACE_TYPE="semantic_highlighter_trace"
+      EXTRA_ARGS=()
+  elif [[ "$TASK_TYPE" == "MetricsCorrelationTrace" ]]; then
+      NOX_TRACE_TYPE="metrics_correlation_trace"
       EXTRA_ARGS=()
   else
       echo "Unknown TASK_TYPE: $TASK_TYPE"